Add Raspistill Manager
Add a raspistill manager. raspistill will take images as fast as
it can and save to a tempfs RAM drive, saving the MicroSD card
from burning out. picture-capturer will be in charge of reading
image thumbnails, detecting motion and moving images for saving
and uploading as needed. Interrupt raspistill at intervals and
restart in order for the camera to re-optimize lighting/settings
on the next set of captures. This improves performance over
taking pictures one at a time, where there's a two+ second delay
when starting up raspistill (not good when taking pics one at
a time).