Detect Motion In Captured Images
Without an added motion sensor to the Raspberry Pi, use the
sequence of captured images from raspistill-manager to
detect motion in motion-capturer.  Move images out of the
temporary RAM drive to the MicroSD card, ready for upload.
Note that uploads are not happening in this commit. Use
image embedded thumbnails placed by raspistill for
comparison vs. full images to use much less CPU with
limited Pi resources. Limit comparison to hotspots; set
box(es) in the image so that very active parts of the image
can be ignored e.g. high traffic road or tree in the wind.