Refactor LIFO 'Queues' To More Proper LIFO Stacks
New images are processed in a LIFO (last-in-first-out) stack
instead of a stack-like queue that didn't really follow a known
queue pattern. Refactor to make it more clear that the latest
image captured will be processed and uploaded first rather than
a FIFO (first-in-first-out) queue, so that users will have access
to the newest images as available.