Post Uploaded Images To Slack Channel As MVP
To get more immediate use of security camera(s), post all uploaded
images to a slack channel called #twin-tiger. Note that an app will
need to be created via the slack website and SLACK_WEBHOOK env var
will need to be set in the Lambda function in the AWS managment
console.